server/proxy: move proxy_settings_mirror to pf_common.c
