+++ server/proxy/pf_common.h
@@ -5,5 +5,6 @@
 #include "pf_context.h"
 
 BOOL pf_common_connection_aborted_by_peer(proxyContext* context);
+void proxy_settings_mirror(rdpSettings* settings, rdpSettings* baseSettings);
 
 #endif /* FREERDP_SERVER_PROXY_PFCOMMON_H */
