+++ server/proxy/pf_context.h
@@ -82,6 +82,5 @@ typedef struct proxy_to_server_context proxyToServerContext;
 
 rdpContext* proxy_to_server_context_create(rdpSettings* baseSettings, char* host,
         DWORD port, char* username, char* password);
-void proxy_settings_mirror(rdpSettings* settings, rdpSettings* baseSettings);
 
 #endif /* FREERDP_SERVER_PROXY_PFCONTEXT_H */
